# Speed up special.ai_zeros/bi_zeros by 10x