Fix performance issues in airyzo fortran code: (i) use only required number of terms in asymptotic expansions, (ii) don't use single-precision arithmetic so that asymptotic root locations are more accurate
```
BENCHMARK                                    BEFORE      AFTER     FACTOR
special.Airy.time_ai_zeros                 474.31ms    43.39ms   0.09147932x
special.Airy.time_bi_zeros                 468.62ms    42.40ms   0.09047139x
```