@raoulbq 2015-03-11
Indeed. However it seems not to affect the runtime of the Hermite computation.
Maybe I compiled it the wrong way.