@ewmoore 2015-03-11
This discussion really belongs in the other PR, but have your profiled it?  Calculating the airy roots happens once.  If I run just `cProfile.run('h_roots_asy(1000)')` on your current code from #4587 (that is not including these changes) I find nearly all of the time in `pbcf`.