@raoulbq 2015-03-11
Right, that explains why I do not see a speedup. The parabolic cylinder function approximation is the most expensive part, hence most time is spent there. This fits.